refactor: Detail 페이지
- getTargetDongil 함수 분리하여 가독성 및 유지보수성 개선
- 추후 백엔드 enum 리팩토링에 따라 추가 개선 진행예정
