+++ src/components/home/pending/PendingDongilItem.tsx
@@ -1,9 +1,9 @@
 import SuggestBadge from '@components/common/badges/SuggestBadge';
 import { modals } from '@components/common/modals/Modals';
 import useModals from '@lib/hooks/useModals';
+import { IDongil } from '@lib/types/IDongil';
 import { EDongilStatus } from '@lib/types/TDongilStatus';
 import { getDate } from '@lib/utils/get/getDate';
-import { IDongil } from '@store/slices/walkingDongilsSlice';
 import { Dispatch, SetStateAction } from 'react';
 import styled from 'styled-components';
