+++ src/components/home/thisWeekS/ThisWeekSDongilList.tsx
@@ -1,5 +1,5 @@
+import { IDongil } from '@lib/types/IDongil';
 import { TItemName } from '@lib/types/TItemName';
-import { IDongil } from '@store/slices/walkingDongilsSlice';
 import { Dispatch, SetStateAction } from 'react';
 import styled from 'styled-components';
 import ThisWeekSDongilItem from './ThisWeekSDongilItem';
