+++ src/lib/mocks/dongils.ts
@@ -1,4 +1,4 @@
-import { IDongil } from '@store/slices/walkingDongilsSlice';
+import { IDongil } from '@lib/types/IDongil';
 
 export const dummyDongils: IDongil[] = [
   {
