+++ src/store/slices/proposedDongilsSlice.tsx
@@ -1,8 +1,8 @@
+import { IDongil } from '@lib/types/IDongil';
 import { TFetchStatus } from '@lib/types/TFetchStatus';
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import { AxiosInstance } from 'axios';
 import { RootState } from '../app/store';
-import { IDongil } from './walkingDongilsSlice';
 
 interface IProposedDongil {
   userName: string;
